requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Faker==25.8.0
ffmpeg-python==0.2.0
filelock==3.15.4
fire==0.7.0
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.53.0
fsspec==2024.6.1
future==1.0.0
huggingface-hub==0.26.2
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai-whisper==20231117
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pdf2docx==0.5.8
pillow==10.3.0
PyMuPDF==1.24.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sentencepiece==0.2.0
setuptools==70.0.0
six==1.16.0
sympy==1.13.1
tbb==2021.12.0
termcolor==2.5.0
tiktoken==0.7.0
tokenizers==0.20.3
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.66.4
transformers==4.46.2
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.1.2
whisper==1.1.10
sacremoses==0.1.1
